Make chain monitoring generic for replay + start with a given tip
This takes the main chain-replay logic out of spawn_chain_monitor
and puts it in its own function, so that we can call it on startup
to replay the chain for ChannelMonitors and ChannelManager. It
further makes spawn_chain_monitor start at a given hash so that
there is no replay between rescan and initial chain monitoring.