Rewrite a multi-threaded daemon with distinct cli binary
This is a meaningful rewrite of the previous sample rust-lightning
based node. Mainly, it switches the code architecture to a new
threading model, where each major LN components gets its own and
communicate with others through dedidcated channels. It also splits
the command line in its own binary, move logging in its own file,
and vets the whole with a configuration file.

Documentation is work in progress and aims to expose Rust-Lightning
architecture usage.

This is still experimental software and shouldn't be used beyond
a local regtest. You still have wide holes which would provoke
certain loss of funds.