Custom Keysmanager and openchannel_without_peer_addr added
MyKeysManager is a custom keysmanager that can be used to set custom channel secrets. Setting custom channel is essential for
interacting with Lnprototest.
A method called openchannel_without_peer_addr is also added. This allows creating a channel with a peer that is already connected. This is also needed for Lnprototest tests regarding open channel.