Stop persisting peer addresses to disk and use the public gossip db
Its always been a bit strange that we persist the addresses of our
peers to disk when there's a public gossip db, its not really a
thing most LN nodes do. It might be useful for private channels but
there's not a lot of reason to have it in `ldk-sample`.